fix W errors

+++ src/app_control.c
@@ -6,7 +6,7 @@
 /*   By: ikarjala <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/10/16 20:41:16 by ikarjala          #+#    #+#             */
-/*   Updated: 2023/01/03 17:37:53 by ikarjala         ###   ########.fr       */
+/*   Updated: 2023/01/03 18:31:23 by ikarjala         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -32,8 +32,8 @@ int	key_axis(int key, const int pval, const int nval)
 
 t_cx	mouse_scale(int x, int y, t_vrect view)
 {
-	const double	w = ft_min(WIN_RESX, WIN_RESY) * 0.5L;
-	const double	amp = 0.01L;
+	const double	w = (double)(ft_min(WIN_RESX, WIN_RESY) * 0.5L);
+	const double	amp = (double)(0.01L);
 
 	return ((t_cx){
 		.x = (double)(x - w) *amp * view.zoom + view.x,
